Update product service for dto, interface and elepment

+++ src/main/java/com/vti/ecommerce/services/IProductService.java
@@ -20,5 +20,5 @@ public interface IProductService extends IBaseService<ProductDto, Long>{
     List<ProductDto> searchProduct(String search, int page, int limit);
     String uploadFile(MultipartFile file) throws IOException;
     //nhập kho:
-    void importProducts(List<ProductImport> productImports, Long supplierId);
+    void importProducts(List<ProductImport> productImports, int supplierId);
 }
